Drafted

+++ CI_Platform1/Models/VolunteeringVM.cs
@@ -32,17 +32,13 @@ namespace CI_Platform1.Models
         public string? Availability { get; set; }
         public string? GoalObjectiveText { get; set; }
 
-
         public string GoalValue { get; set; } = null!;
         public string UserPrevRating { get; set; }
-
-
-
 
         // ..............comment
         public List<Comment> comments  { get; set; }
 
-        public int user_id { get; set; }
+        public long user_id { get; set; }
         public int mission_id { get; set; }
         public string content { get; set; }
         public DateTime created_at { get; set; }
